Initial Commit
This is the start of Milestone 2. I pulled the class from the previous
milestone into the Qt project. I have not yet added the functionality in
for the Image overlay.

Beacuse Qt project files are portable we do not need to set up local
projects, just we cannot merge them so if a Project file change is made
it must be done on the mater branch. IE adding new files and forms.